// Neptune.Web/src/app/shared/generated/model/source-control-bmp-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class SourceControlBMPUpsertDto { 
    SourceControlBMPID?: number | null;
    SourceControlBMPAttributeCategoryName?: string | null;
    SourceControlBMPAttributeCategoryID?: number;
    SourceControlBMPAttributeID?: number;
    SourceControlBMPAttributeName?: string | null;
    IsPresent?: boolean | null;
    SourceControlBMPNote?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface SourceControlBMPUpsertDtoForm { 
    SourceControlBMPID?: FormControl<number>;
    SourceControlBMPAttributeCategoryName?: FormControl<string>;
    SourceControlBMPAttributeCategoryID?: FormControl<number>;
    SourceControlBMPAttributeID?: FormControl<number>;
    SourceControlBMPAttributeName?: FormControl<string>;
    IsPresent?: FormControl<boolean>;
    SourceControlBMPNote?: FormControl<string>;
}

export class SourceControlBMPUpsertDtoFormControls { 
    public static SourceControlBMPID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SourceControlBMPAttributeCategoryName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SourceControlBMPAttributeCategoryID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SourceControlBMPAttributeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SourceControlBMPAttributeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsPresent = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SourceControlBMPNote = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}